# Optionally use fpng encoder for much faster movie writing