--- a/drivers/png/png_driver_common.cpp
+++ b/drivers/png/png_driver_common.cpp
@@ -35,6 +35,12 @@
 #include <png.h>
 #include <string.h>
 
+#ifdef FPNG_ENABLED
+// Must be here since it uses a namespace `fpng`,
+// which must not be nested underneath PNGDriverCommon.
+#include <fpng.h>
+#endif
+
 namespace PNGDriverCommon {
 
 // Print any warnings.
@@ -124,6 +130,36 @@ Error png_to_image(const uint8_t *p_source, size_t p_size, bool p_force_linear, 
 	return OK;
 }
 
+#ifdef FPNG_ENABLED
+// NB: must be called with a pre-decompressed image that is in a format supported by fpng.
+Error _image_to_png_fpng(const Ref<Image> &source_image, Vector<uint8_t> &p_buffer) {
+	static bool fpng_init_called = false;
+	ERR_FAIL_COND_V(!(source_image->get_format() == Image::FORMAT_RGB8 || source_image->get_format() == Image::FORMAT_RGBA8), FAILED);
+	const Vector<uint8_t> image_data = source_image->get_data();
+	std::vector<uint8_t> out_buf;
+	if (!fpng_init_called) {
+		fpng::fpng_init();
+		fpng_init_called = true;
+	}
+	int success = fpng::fpng_encode_image_to_memory(
+			image_data.ptr(),
+			source_image->get_width(),
+			source_image->get_height(),
+			source_image->get_format() == Image::FORMAT_RGB8 ? 3 : 4,
+			out_buf,
+			0 // flags
+	);
+	if (!success) {
+		ERR_FAIL_V_MSG(FAILED, "Failed to encode image to PNG.");
+	}
+	auto bs = p_buffer.size();
+	Error err = p_buffer.resize(bs + out_buf.size());
+	ERR_FAIL_COND_V(err, err);
+	memcpy(p_buffer.ptrw() + bs, out_buf.data(), out_buf.size());
+	return OK;
+}
+#endif
+
 Error image_to_png(const Ref<Image> &p_image, Vector<uint8_t> &p_buffer) {
 	Ref<Image> source_image = p_image->duplicate();
 
@@ -132,6 +168,13 @@ Error image_to_png(const Ref<Image> &p_image, Vector<uint8_t> &p_buffer) {
 	}
 
 	ERR_FAIL_COND_V(source_image->is_compressed(), FAILED);
+
+#ifdef FPNG_ENABLED
+	if ((source_image->get_format() == Image::FORMAT_RGB8 || source_image->get_format() == Image::FORMAT_RGBA8)) {
+		// Format supported by fpng, use it.
+		return _image_to_png_fpng(source_image, p_buffer);
+	}
+#endif
 
 	png_image png_img;
 	memset(&png_img, 0, sizeof(png_img));