… thirdparty/fpng/fpng.cpp