This PR proposes a compile-time flag to use [fpng](https://github.com/richgel999/fpng/) for PNG encoding instead of libpng.

On my machine (2023 Macbook Pro, M2 Max, 64 GB memory, macOS 15.2), there's a **15-fold** difference in rendering speed for movie maker mode, and I'm now able to write these videos at almost realtime 50 FPS speed.

The downside is that `fpng` only supports encoding RGB and RGBA frames, which is why this is disabled by default.

### No fpng (`scons dev_mode=yes no_editor_splash=yes fpng=no`) 

```
Godot Engine v4.4.dev.custom_build.f9f599aa5 (2025-01-08 07:22:55 UTC) - https://godotengine.org
Metal 3.2 - Forward+ - Using Device #0: Apple - Apple M2 Max (Apple8)

Movie Maker mode enabled, recording movie at 50 FPS...
2025-01-08 09:35:02.707 godot.macos.editor.arm64.nofpng[32266:3401569] +[IMKClient subclass]: chose IMKClient_Modern
2025-01-08 09:35:04.456 godot.macos.editor.arm64.nofpng[32266:3401569] +[IMKInputSession subclass]: chose IMKInputSession_Modern
----------------
Done recording movie at path: /.../XX/frame.png
451 frames at 50 FPS (movie length: 00.0:00.0:09.0), recorded in 00.0:02.0:15.0 (6% of real-time speed).
CPU time: 0.03 seconds (average: 0.06 ms/frame)
GPU time: 0.00 seconds (average: 0.00 ms/frame)
```

### Yes fpng (`scons dev_mode=yes no_editor_splash=yes fpng=yes`) 

```
Godot Engine v4.4.dev.custom_build.3d0cb67ad (2025-01-08 06:58:20 UTC) - https://godotengine.org
Metal 3.2 - Forward+ - Using Device #0: Apple - Apple M2 Max (Apple8)

Movie Maker mode enabled, recording movie at 50 FPS...
2025-01-08 09:39:39.064 godot.macos.editor.arm64.fpng[33081:3415884] +[IMKClient subclass]: chose IMKClient_Modern
2025-01-08 09:39:40.891 godot.macos.editor.arm64.fpng[33081:3415884] +[IMKInputSession subclass]: chose IMKInputSession_Modern
----------------
Done recording movie at path: /.../XX/frame.png
451 frames at 50 FPS (movie length: 00.0:00.0:09.0), recorded in 00.0:00.0:10.0 (90% of real-time speed).
CPU time: 0.02 seconds (average: 0.04 ms/frame)
GPU time: 0.00 seconds (average: 0.00 ms/frame)
----------------
```