@akx 2025-01-08
> What are the negatives of using fpng? Compatibility?

As said in the description, fpng only supports encoding RGB and RGBA frames (libpng does grayscale images too, `FORMAT_L8` and `FORMAT_LA8` in Godot nomenclature). fpng also generates larger files (1.69x the size of libpng for my example above), but that probably isn't a big deal.

fpng's readme says "The files written by fpng conform to the PNG standard, are readable using any PNG decoder [...]", so compatibility is not a concern.