@akx 2025-01-08
@Chaosus btw: `topic:import` should probably be `topic:export`.

fpng does have a decompressor too ("2.5-3x faster decompression"), but this doesn't enable it :)