@Calinou 2025-01-18
Can you benchmark https://github.com/godotengine/godot/pull/80768 on the same machine to see which one is faster (QOI versus fpng)? Recent FFmpeg versions accept QOI as input files.