@akx 2025-01-21
> Can you benchmark #80768 on the same machine to see which one is faster (QOI versus fpng)? Recent FFmpeg versions accept QOI as input files. Reading this PR description, it seems fpng is quite a bit faster than QOI, so I'd be interested to know about file sizes too.

Um, AISI that PR doesn't have QOI _saving_ at all, so it doesn't apply for this (as mentioned in my comment https://github.com/godotengine/godot/pull/101268#issuecomment-2577106974 this doesn't touch importing).