@akx 2025-01-21
> I get a build error on MSVC 2022 with `target=template_release fpng=yes`:

I'll check that out – but first I'll have to figure out why this just stopped building altogether on my system 😩