@akx 2025-01-21
Rebased.

@Calinou Fixed the build on my end (but I don't readily have a Windows machine to try your scenario with).

@bruvzg Done. If the feature is enabled, it is now used for RGB/RGBA encoding, and everything else is passed through to libpng.