@akx 2025-01-21
> Can you benchmark #80768 on the same machine to see which one is faster (QOI versus fpng)? Recent FFmpeg versions accept QOI as input files. Reading this PR description, it seems fpng is quite a bit faster than QOI, so I'd be interested to know about file sizes too.

Ah – I found #91263 which does exporting too and tried with that.

* QOI is faster than fpng (but of course is not PNG), does realtime encoding on my machine.
* QOI is pretty much the same size as libpng.
* `ffmpeg -f image2 -framerate 50 -i foo%08d.qoi -pix_fmt yuv420p -tune animation -crf 15 ...` works fine for encoding a QOI sequence.

```
$ du -hs tmp-*
940M    tmp-fpng
573M    tmp-libpng
576M    tmp-qoi
```

## QOI via #91263

```
$ ../godot.macos.editor.arm64.qoi --write-movie ../tmp-qoi/foo.qoi --fixed-fps 50 --audio-driver Dummy --minimized amv.tscn
Godot Engine v4.4.dev.custom_build.bd7d2cccf (2025-01-01 22:45:50 UTC) - https://godotengine.org
Metal 3.2 - Forward+ - Using Device #0: Apple - Apple M2 Max (Apple8)
451 frames at 50 FPS (movie length: 00.0:00.0:09.0), recorded in 00.0:00.0:09.0 (100% of real-time speed).
CPU time: 0.02 seconds (average: 0.04 ms/frame)
GPU time: 0.00 seconds (average: 0.00 ms/frame)
```

## fpng via this PR

```
$ ../godot.macos.editor.arm64.fpng --write-movie ../tmp-fpng/foo.png --fixed-fps 50 --audio-driver Dummy --minimized amv.tscn
Godot Engine v4.4.beta.custom_build.9eacdcdeb (2025-01-21 13:13:20 UTC) - https://godotengine.org
Metal 3.2 - Forward+ - Using Device #0: Apple - Apple M2 Max (Apple8)
451 frames at 50 FPS (movie length: 00.0:00.0:09.0), recorded in 00.0:00.0:10.0 (90% of real-time speed).
CPU time: 0.02 seconds (average: 0.04 ms/frame)
GPU time: 0.00 seconds (average: 0.00 ms/frame)
```

## libpng

```
$ ../godot.macos.editor.arm64.nofpng --write-movie ../tmp-libpng/foo.png --fixed-fps 50 --audio-driver Dummy --minimized amv.tscn
Godot Engine v4.4.beta.custom_build.9eacdcdeb (2025-01-21 13:13:20 UTC) - https://godotengine.org
Metal 3.2 - Forward+ - Using Device #0: Apple - Apple M2 Max (Apple8)
451 frames at 50 FPS (movie length: 00.0:00.0:09.0), recorded in 00.0:02.0:27.0 (6% of real-time speed).
CPU time: 0.03 seconds (average: 0.06 ms/frame)
GPU time: 0.00 seconds (average: 0.00 ms/frame)
```