@akx 2025-08-08
> The fpng encoder encodes in a popular open standard unlike qoi which makes this attractive. (citations of popularity of qoi welcome)

QOI is actually a ~quite OK~ very decent format for movie-maker purposes (and worked great for me, see comment above). It's been nicely supported by ffmpeg for 3 years now (since https://github.com/FFmpeg/FFmpeg/commit/973fab565378cbdd0712977152a66f5b17938d51), so I would like to see support for both it and fpng-based encoding in Godot :)